Be more careful about parsing Olsrd Route Commands
This caused a HUGE problem! Without this check,
there were cases where the underlying code would
attempt a DNS resolution on non-ip addresses. This
caused the handler thread to wait and caused
everything to get backed up. It looked like it was
the thread blocking/buffering input from the child
process, but that was not the case!